refactor: add context, divide hooks, folder structure
