+++ public/data/product_1.json
@@ -1,7 +1,11 @@
 {
-  "price": "150000",
+  "name": "조던 1 미드 라이트 스모크 그레이",
   "category": "shoes",
-  "image": "https://kream-phinf.pstatic.net/MjAyMTA3MjhfMjIg/MDAxNjI3NDQxMDA1NjE5.HOgIYywGZaaBJDqUzx2OnX9HAxoOWPvuWHqUn_LZGcgg.VYIuOfA5_GgjBGRowv6dmQuAOPtUvmAxbGpOyUCOCtYg.PNG/p_9d8ed1a74d2540ab9842e63363607bf4.png?type=m",
-  "brand": "Nike",
-  "name": "나이키 덩크 로우 레트로 블랙"
+  "id": "f5fef828-1d00-4bde-989c-58eddcf8cd12",
+  "brand": "Jordan",
+  "createdAt": {
+    "seconds": 1675175364,
+    "nanoseconds": 66000000
+  },
+  "price": 186000
 }
