+++ src/features/cart/index.ts
@@ -1,2 +1,4 @@
 export * from "./cart-item";
 export * from "./cart-list";
+export * from "./hooks";
+export * from "./service";
