+++ src/features/cart/service/index.ts
@@ -1,0 +1,2 @@
+export * from "./cart.client.service";
+export * from "./cart.model";
